feat: threadContainer
